requirements.txt
# Core
Django==3.0.3
gunicorn==20.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.2.1
psycopg2-binary==2.8.4

# Django addons
django-filter==2.2.0
django-rest-passwordreset==1.1.0
django-admin-rangefilter==0.6.3
django-storages==1.9.1
django-heroku==0.3.1
python-decouple==3.3
python-dateutil==2.8.1
drf-yasg==1.17.1

# Celery
celery==4.4.2
redis==3.4.1
django-celery-beat==2.0.0
django-celery-results==1.2.1

# AWS S3 storage
boto3==1.12.30
botocore==1.15.30

# Google services
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gspread==3.6.0
oauth2client==4.1.3
oauthlib==3.1.0
firebase-admin==4.0.0

# General libs
Pillow==7.0.0
Markdown==3.2
requests==2.22.0
Unipath==1.1
whitenoise==5.2.0

# IPython
ipython==7.18.1
jedi==0.17.2
parso==0.7.1
